fix KeyNotFoundException that could occur in SearchInCity()
Sometimes spravnik.com redirects from city page back to its province page.
In such cases it is impossible to search anything in the city, so the only
thing we could to is to skip it.

Previously code didn't handle such sutiation and also selected any <form> element
on page to do search. Turns out that province page also contains form (for search cities
inside it), so when request to city URL was redirected back to province URL, code
didn't notice it and tried to submit search to that form. As form on the province page
doesn't have 'soname' and 'io' fields, KeyNotFound exception was thrown by form.SubmitAsync()

Now code uses more specific selector to find search form: 'serchForm form' (sic)
This way, it the site redirects back to province page, search form is not found
and SearchInCity() returns empty sequence